PCHR-626: Add blocks box-shadow
.panel and .page-header have now box-shadow so they stand out more from the background (the property value is standardized in the $box-shadow property)

I had to implement a small hack for the contextual variations of .panel (.panel-primary, .panel-info, etc). Given that in the variations .panel-header is detached from the rest of the panel, the box-shadow cannot be applied to the whole .panel, instead it must be applied individually to .panel-header, .panel-body and .panel-footer. Being -body and -footer attached, the latter's shadow overflows into the former. Since we must stick with bootstrap markup (and as such I couldn't add a simple wrapper to -body and -footer to which i could've applied a box-shadow to), I had to resort to a .panel-body:before pseudo-element with the same background color as the panel-body bg color, which results in the overflowing shadow being hidden