PCHR-927: Create AbsenceType entity
This is the basic structure of the AbsenceType entity. At this point,
it only has the create method, which can be used to create and
update Absence types. The method runs a large number of
validations to make sure only types folowing all the rules are
stored in the database.