PCHR-902: Notify the user about AbsenceType BAO validation errors
The AbsenceType BAO has a lot a validations (more than those made on the
form). That is because it may be used not only by forms, but alsi by APIs,
scripts and anywhere else. So we try to make everything possible to
prevent the creation of invalid types.

Since these are business rules validations and they throw exceptions, we
can't use they in the form validation process, so we show any error as a
flash message to the user.

On the form side we also do a bunch of validations, but there we are concerned only
on checking if the form if properly filled (like if required fields are not empty, numeric
fields have numbers, both expiration fields are not empty). If a mailicious user try
to hack it (for example, setting the value of a hidden field), the AbsenceType validation
will get it and the new type won't be created.