PCHR-902: Make sure that the Absence Types admin menu always point to the list page
Civi StateMachine stores the current controller action on the session and, if it is not
reset and the page is accessed without an explicit action, the stored one will be used.
That means that if you were viewing the edit form and tried to go back to the list through
the admin menu, you were end up on the edit form again (because that was the action stored in
the session). To solve this, I changed the menu URL to explicit include the browse action.