PCHR-902: Minor adjustment of how TOIL and Carry Forward expiration fields are displayed
With this change, the expiration duration and date fields will be visible
after  the page load if only one of its fields was filled. This is important,
because there's a form rule that validates if the expiration duration or date
is incomplete and if the fields are hidden, the use will not see the error.