PCHR-907: Add findWithNumberOfWeeksAndHours method to WorkPattern BAO
This is method wraps the find method, joining the WorkPattern with
WorkWeeks and WorkDays so we can get the number_of_weeks and
number_of_hours for each pattern.

For the joins to work, the links method was added to every entity
used by Work Patterns. This method is used by the DAO::joinAdd
method and it returns an array describing how the entities relate
to each other.