PCHR-946: Add the Work Pattern form with Detail tab
This is the first part of the Work Pattern form. At this point,
it only contains the Detail tab form. There's a delete button
on the form, but it's action will only be implemented with the
PR that will add the form for the Calendar tab.

A small change was made the AbsenceType form so we
could use the same CSS rules to position the submit buttons
on both forms.