PCHR-947: Add the Calendar tab to the Work Pattern form
The Calendar tab shows the work pattern configuration. At this point, it can only show the
data, but it's not possible to update it yet.

The fields names are generated in the format weeks[weekIndex][days][dayIndex][fieldname],
to match the structure of the defaultValues loaded from the WorkDay::getArrayValues
method.