PCHR-947: Change WorkDay time_from and time_to type to CHAR
The DataObject class, which is used by CiviCRM to save records to the database,
can't handle fields of the Time type. When updating or inserting new entities with
fields of this type, the given value was converted to an int (Civi does this for every
type it can't handle) making the wrong value to be stored.

On CiviCRM, when there's a datetime field, the date needs to be converted to a
timestamp before storing it to the databse. We can't do the same to a time field,
and since we won't need any ordering or calculation of these values, storing they
as a string is fine.