PCHR-947: Add the possibility to show/hide weeks according to the selected number of weeks
There's a hidden field for each week on the form that is used to keep track if
its respective week is visible (a value of 1 means visible, 0 means hidden).

When opening the form, we add these values dinamically to the defaultValues
property, so they can be used to set the flag fields values.