PCHR-947: Add logic to reset and set the initial values of the Calendar form
When adding a new Work Pattern, the calendar should display a single week
with 5 working days and 2 days as weekend. The times, break and leave days
fields of the weekend should be disabled.

When editing an existing Work Pattern, the "number of weeks" field can be
used to change the number of weeks visible. When used to increase the
number of items, the new visible weeks will have the same default values
as when adding a new Work Pattern. When used to decrease the number of
weeks, the "removed" weeks are hidden and all of its fields are reset.