PCHR-947: Change the calendar form to save the weeks and days for a Pattern
The WorkPattern and WorkWeek were changed so they could receive the
weeks and days from the form and save it.

On the WorkPattern class, the create method now is able to receive a 'weeks'
array inside the $params, containing the weeks and days to be added to the
pattern.

On the WorkWeek class, the create method now is able to receive a 'days'
array inside the $params, containing the days to be added to the week.