PCHR-947: Check if properties exist when validating a Non Working Day
Non Working Day fields are disabled and, in being so, they are not sent when
the form gets submitted. Because of that, in a perfect scenario, they are not
available to the form, and trying to access its times or break will result
in a trying to access an undefined index error. To solve this, the validation
code now checks is the index is defined.