PCHR-947: Properly display the number of hours for each week
When opening the WorkPattern form, the number of hours for each
week is calculated based on the loaded default values. If the form
has been submitted, it calculates the number of hours based on the
submitted values.

Also, whenever the user updates the times or break for a day, the
number of hours for that week will be automatically updated to reflect
the changes.