PCHR-947: Add input mask to the times and break fields of the Calendar form
To the "Time From" and "Time To" mask, it was added a mask with the 99:99 format.
To the "Break" field, it was added a mask of the decimal format, that allows only numbers
and a decimal separator.

After adding the mask to the fields, the input event was triggered, so I used the oncomplete
event of the mask to update the number of hours for the day and the week.

There's also a change in the CSS to align the fields text to the center.