PCHR-947: Change the delete button of the Work Pattern form work
When the user clicks on the "Delete" button, a confirmation is displayed and,
if they confirm it, the browser is redirected to the "delete" action, which will
call the WorkPattern BAO del method to delete the pattern.

In the future, when we need a more complicated behavior like not deleting
work patterns already assigned to employees, all we'll have to do is add this
validation to the BAO's del method.