PCHR-947: Refactor the javascript code of the WorkPattern form
The code was extracted to a separate js file and was split into three classes.
The idea was to better separate the code that deals with the form in general,
the one that only deals with the week and the one that deals with days of the week.

I've also added a few missing docblocks to some methods of the WorkPattern form class.