PCHR-1119: Add all the L&A entities to the hook_civicrm_entityTypes hook
The hook only added the AbsenceType entity to the list of available entities and,
though it worked on CiviCRM 4.5, on 4.7 it caused a lot of problems with any code that
used the API for any of the missing entities.

The "Delete", "Enable", "Disable" and "Set as default" actions of the Work Patterns list,
all use the Work Pattern API to update the selected item. After moving to 4.7 they all
stopped working. This change fixed all of the errors.