PCHR-1119: Fix L&A tests to run on CiviCRM 4.7
The Unit tests were changed to implement the HeadlessInterface and its
setUpHeadless method to automatically install the extension on the
test environment.

For the WebTests I just removed an unecessary require statement from the
test classes, as we don't need they with the new phpunit bootstrap that uses
cv and Civi's class autoloader.

I also made a small change to the FormTest of the AbsenceType. After the
update, the isVisible method was failing to check for select2 elements, so I
needed to add th s2id_ prefix to the field name to make it work.