PCHR-952: Create the AbsencePeriod Entity
The AbsencePeriod is just a simple class to represent a period of time.
It has the start_date and end_date fields, which both are mandatory
and the start_date can't be greater or equal to the end_date.

The weight is used to determine which is the previous/next period. If
someone tries to create a period without the date, it will automatically be
the maximum existing weight + 1. If someone tries to insert or update a
period with an weight already in use, all the period with the weight equal
to or greater than it will be increased.