PCHR-889: Fix the search description for Job Roles regions on the Adv. Search
When searching for Job Role Regions, the search values where not displayed on the
search description. For example, when searching for the "North" region, the user would see this:

Job Role Job Role Region IN "" ...AND...

instead, they should see this:

Job Role Job Role Region IN "North" ...AND...

This was happening because instead of the Region Options, we were passing the
list of Department Options to the method that builds the description making impossible
to get the correct name of the selected region.