PCHR-889: Add the pseudoconstant option to the HrJobRoles DAO fields
The HrJobRoles extension was extracted from the HrJobContract extension,
but when this extraction was made the new HrJobRole DAO was copied without
the pseudoconstant option for the location, region, department and leve type
fields. For this reason, the Hrjobroles_BAO_Query class was referencing the
hrjobcontract classes to load the options values.

Since these classes should be removed in the future
(https://compucorp.atlassian.net/browse/PCHR-1026), this change already
eliminates this dependency on them.