PCHR-1110: Add the getNumberOfWorkingDaysToWork to the AbsencePeriod class
This method returns the number of days to work on an AbsencePeriod between the
given start and end dates.

This method doesn't count days outside the AbsencePeriod. Meaning that, If the given
start date is less than the AbsencePeriod start date, then the AbsencePeriod's start
date will be used. If the given end date is greater than AbsencePeriod end date,
then the AbsencePeriod's end date will be used.

Also, a small check was added to the getNumberOfWorkingDays method to only
allow it to be called on an AbsencePeriod with valid start and end dates.