PCHR-1060: Minor refactoring and adding helpers methods to EntitlementCalculation
Part of the getPreviousPeriodProposedEntitlement was extracted to the
getPreviousPeriodEntitlement, which is returns the Entitlement instance.

Also, getPreviousPeriodProposedEntitlement is now public, so that it can
be used to help populate the Entitlement Calculation page in the future.
The getNumberOfLeavesTakenOnThePreviousPeriod was also made public
for the same reason.

Still thinking about the date that will be needed for the  Entitlement Calculation
page, a new helper method was added: getNumberOfDaysRemainingInThePreviousPeriod

Finally, getJobLeaveForAbsenceType and getContractDetails were made
private, as this is how they should have been from start.