PCHR-1039: Add the getActiveContracts to the HRJobContract BAO
This method returns all the active contracts between a startDate and
an endDate. If no dates are given, the current date will be used instead.

The hierarchy of JobContract class and its related entities is a little bit
complicated and without this method, one would have to take care of a
lot of things like checking the latest contract revision and looking for date
on the jobdetails entity.

A new API action was added as well. It has the same name and basically
just relay the job to the BAO method.