PCHR-1210: Refactor AbsenceType BAO tests to run faster
The dependency on CiviUnitTestCase was making the tests perfomance
very poor. By removing it, the average time to run the tests decreased from
31s to 1.8s \o/

In order to remove CiviUnitTestCase, a small change to the createReservedType
method was required. It was using the getConnection() method to get a
connection witht the test database and insert data on it. I replaced this with
the CRM_Core_DAO and its executeQuery method.