PCHR-1039: Fix the hrjobcontract ext installation for the EntitlementCalculationTest
The hrjobcontract has a lot of upgrade functions, but the install() method from CiviEnvBuilder
is not capable of running them automatically. Because of that, if the test relies on a table or
field that is created during the upgrade, it will fail. To solve this, we instantiate the
extension's upgrader and call it's install method.

The upgrade function upgrade_1015 was missing on the Upgrader.upgradeBundle method,
so it was added in order to get the extension properly upgraded.