PCHR-1039: Add the ManageEntitlements page
This page receives the ID of an AbsencePeriod and display a list of EntitlementCalculations
for all the active contracts during the given period. There are multiple entries for the same
contract, one for each enabled AbsenceType.

It should be possible for the user to override the calculated proposed entitlement.
For this, it was added an edit button and an override checked next to the proposed
entitlement value. When clicked, it hides the value and display an edit field where
the user can input the new value.

At this point we only have the fields and javascript code to hide and display them, but
the value isn't stored anywhere, as this will be implemented on another PR.