PCHR-1087: Fix two AbsenceType BAO tests that required an specific database state
Both The testShouldNotAllowColorToBeReusedUntilAllColorsHaveBeenUsed and the
testShouldHaveAllTheColorsAvailableIfTheresNotTypeCreated expect the database to
have no existing absence types, but when the extension is installed by the setupHeadless
method, the system default types are created, making the tests fail.

Those tests worked before because we used the CiviUnitTestCase that would truncate the
table after each test, but now that this dependency was removed, they started to fail. For
some reason, I just noticed this now :(