PCHR-1088: Add the "Export to CSV" action to the Entitlement Calculations page
This is a link at the top right side of the entitlement calculations list and, when clicked,
exports a CSV file with exactly the same information displayed on the screen.

As we need to export the data on the screen, even if the user has overridden some of
the entitlements, the file is generated by submitting the form so we can include the proper
values in the file.