PCHR-1089: Add the "Add 1 Day" and "Copy to All' actions buttons
These buttons are inside the "New Proposed Entitlement" header, and work as follow:

Add 1 Day button: Adds an extra day to every entitlement on the list and marks the
entitlement as overridden.

Copy to All: Copy the value of the first proposed entitlement on the list to all others and
marks the entitlements as overriden.