PCHR-1108: Performance inprovements for the entitlement calculation
Two improvements were made in order to achieve a better performance during the
entitlement calculation:

- Cache the result of AbsencePeriod.getNumberOfWorkingDays(). The same AbsencePeriod
instance is used for every calculation of a given period, so, by caching this, we avoid a lot of
queries to get the number of public holidays and a loop through all the days in the period.

- The chained API call in ManageEntitlements.getActiveContractsForPeriod() was resulting in
a lot of queries to get the related contacts. Now we use the HRJobContract directly (to get
rid of the API overhead) and get all the contacts with a single query/API call