PCHR-1114: Fix the submission of the entitlement calculation form
The Form class will only process a form submission if it was submitted
by a button added to the form with the $this-addButtons() method. As
there was no button added to this form, the postProcess() method was
never called.

The button that was "hardcorded" in the template file was removed
and is now rendered from the form buttons, using the
CRM/common/formButtons.tpl template (the same one used by other
forms in this extension)