PCHR-1114: Change the ManageEntitlements form to be able to return to an URL after saving entitlements
Before this change, after saving the Entitlements the user was
redirected back to the calculations page. Now, when accessing
the form, coming from another page, we store the referer URL
in the session. This way, after saving the entitlemens, we can
redirect the user back to where they came from.

If the the calculations page was accessed directly, the old it
will work just like before, and the user will be redirected to the
calculations page again.