PCHR-1060: Performance improvements for the entitlement calculation
A couple of improvements was made:

- AbsencePeriod.getPreviousPeriod: The same AbsencePeriod instances is used to
every calculation. By caching the return of this method, we can drastically reduce the
number of database queries.

- PublicHoliday.getNumberOfPublicHolidaysForPeriod. This method is called multiple
times during the entitlement calculation and, most of the time, with the same parameters.
To avoid repeated database queries, we cache the returned value for each given set of
parameters. A method to clear the cache was created as well.