PCHR-1064: Redirect to the Manage Entitlements page on contract save
When adding or editing a contract, a confirmation message will be displayed
saying that, if the user proceeds, the leave entitlements will be updated.
This update happens on the Manage Entitlements page, meaning that,
after saving the contract, the user will be redirected to it. The ID of the
contract's contact is added to the URL as the cid parameter in order to
only update that employee entitlements.

To make it easier the redirect to the Manage Entitlements page, the
mandatory paramenter id is now optional. When not passed in the URL,
it will use the current period by default.