PCHR-1066: Allow AbsenceType with add_public_holiday_to_entitlement to be updated
Trying to update an existing Absence Type where add_public_holiday_to_entitlement
was checked, was throwing the error: There is already one Absence Type where "Must
staff take public holiday as leave" is selected.

That happened because the validation rule that checks for existing Absence Types
with add_public_holiday_to_entitlement == 1 wasn't taking into account if the existing
type was the one being updated. I fixed this, by adding a "id <> $this->id" (when
the $type id is available) condition to the validation query, so it won't consider the type
being updated as the existing one.

I've also changed the validation to use the DAO to executed the query instead of the API.
There's no reason for the extension to use it's own API.