PCHR-1146: Set the default phone location type when creating/editing a contact
When creating a new contact, the location type of the first phone number
is automatically set to the default location type, but this doesn't happen to
the other numbers, which end up with another location type by default.

We use the buildForm hook to alter the form in order to change the default
value of the location type field to the default location type. This is done only
if the phone number empty, so to avoid overwrite any previously selected
value.