PCHR-914: Refactor WorkPattern web tests
Extracted the methods that handle the form to a trait and left the test class
with only test methods. I did this in order to be able to reuse the trait methods
in other tests. For example, they be useful to create a work pattern for a test of
the list page.