PCHR-914: Refactor WorkPattern BAO tests to not depend on CiviUnitTestCase
Instead of using the API to create WorkWeeks and WorkDays, we're now using
the BAO create method. There's no reason to use the API if the classes being
used are part of the same extension.

By doing this, we're don't need the callAPISucess and CiviUnitTestCase anymore,
and the tests are much faster now! \o/