PCHR-1343: Refactor the code that install option groups from XML files
An install method was added to the extension Upgrader
class. This extends the logic of the Base Upgrader to
search for _install.xml files in other directories.
Currently, the implementation only looks for files
inside the xml/option_groups dir.

With this change, we can now have each option group
definition in a separate file, improving the code
organization. The auto_install.xml file was removed and
its contents were split into two new files in xml/option_groups.