PCHR-1347: Add the LeaveBalanceChange entity
This entity is used to represent balance changes to a LeavePeriodEntitlement.
That is, days added or removed from the entitlement. An added day, will be
represented by a LeaveBalanceChange with a positive amount value. Deducted
days are represented as a negative amount. The amount can be fractioned, like
0.5 (half a day).

Some balances change can expire. Those have an expire date and, in a future
PR, a scheduled job will be added to check this date and record a new, negative
change to what has expired.

Some balances can also have a source (so we can track down what caused the
balance change). The idea is to have multiple sources, that's why source_id is not
a real FK, althought in this first version the only existing source will be a leave request
date (still not implemented). In the future, when more sources exist, we'll also need
a source_type field to identify the source.