PCHR-1060: Take LeaveRequests into account on the EntitlementCalculation
Another step to move things to the new data structure for period entitlements and
leave requests.

Before this, EntitlementCalculation had a method to return the number of leaves
taken during the previous period, but it was hardcoded to always return 0, since
leaves requests weren't yet implement. Now, the LeavePeriodEntitlement and
LeaveBalanceChange classes were updated with methods that can retrieve this
information (The number of leaves taken is basically the sum of all balance changes
caused by a leave request for a period entitlement)