PCHR-1351: Completely replace Entitlement with LeavePeriodEntitlement
This is the final step towards using the new data structure for leave period
entitlements and leave requests.

The LeavePeriodEntitlement was updated with a new method, named
saveFromCalculation(), that can be used to save period entitlements (
including all the balance changes and leave requests required) from an
EntitlementCalculation instance. This new method is used mainly by
the Manage Entitlements page, when the user clicks on "Save new
entitlements".