PCHR-1443: Small refactoring to improve the tests execution
There were two different versions of the setUpHeadless method among the test
classes. One would just install the L&A extension and the other would install both
the L&A and hrjobcontract extensions.

Due to this difference, when trying to execute  all the tests at once
(phpunit4 tests/phpunit/CRM), one would see a lot of "Installing test database"
messages on the output. That happens because everytime CiviEnvBuilder find
a setupHeadless method differente from the last one executed, it starts to set up
everything again. To fix this, I created a BaseTest class, which has the most complete
version of setupHeadless method, and made all the other tests classes extend it.