PCHR-1081: Fix DB Error while accessing the Staff Directory page
The code was trying to add the Job Summary custom group fields to
the query which fetch the data for this page, but these fields were
removed months ago by PCHR-774. This lead to the malformed SQL
query that was resulting on the DB Error.

This was fixed by removing the statements which added those fields
to the query.