PCHR-1081: Displays the managers on the staff directory
The Profile was trying to fetch the Manager information from
the jobcontract_role table, but it's not stored there anymore
(neither on jobroles). To get such information, we need to
fetch all the related contacts through the "Line Manager Is"
relationship. This also means a contact can have multiple
managers.

I created a new Query Object to the hrprofile extension that
injects all the necessary joins and fields to fetch the
managers for a contract. Since we have multiple managers
from a single contact and we need to return this information
in a single field, I used the GROUP_CONCAT function to group
all the managers into a single field, separated by commas.