PCHR-1485: Display a job role with no end date as current role
The code wich separates past and current job roles was not dealing
properly roles with empty end dates (that is, an end_date equals to
null or '0000-00-00');

The code was checking for an empty end_date using !end_date, but,
since the end_date was always converted to a moment object, this
condition would always evaluate to false. This has been fixed by
only converting the date to an object if it's not empty.