PCHR-1485: Change type of HrJobRole's start_date and end_date to datetime
In MySQL, when you set a TIMESTAMP field to null it will be automatically updated
to the current date. This behavior was causing problems when an user tried to
update a job role with an empty end_date. Changing the type of the fields to
DATETIME fixed this, since this type allows null values. The default value was also
updated to be NULL instead of 0000-00-00.

There was also a problem where the upgrade method 1003 was never going to be
executed. I bumped the number of the upgrade method for force it to be executed.
The reason for this is that, when CiviCRM installs an extension, it doesn't sets
the upgrade version as the latest available without running the upgrade methods.
So, any installation made after the 1003 method was introduced, never had this
method executed and the field type would still be "TIMESTAMP".