PCHR-1602: Allow getting Active Contracts for an specific contact
This changes the HRJobContract::getActiveContract method to
allow fetching contracts only for a specific contact.

Another change introduced is that now the method returns the
contract details. Hence it was renamed to getActiveContractsWithDetails.
This change is a preparation for other future changes on the Entitlement
Calculation, where we'll be dealing most with the contract dates and
having it without an additional query is important.

Finally, the method was also changed to return the contracts ordered by
the start date, so one can be confident about what to expect in terms of
order