PCHR-1603: Add the getContactsWithActiveContracts method to HRJobContract
This method is similar to the getActiveContractsWithDetails, but instead of returning a
list of active contracts it returns a list of contacts which has at least one active contract
during the given period.  If no period is given, then it returns contacts with active
contracts on the current date.

An API endpoint was also added to enable this to be called by other extensions