PCHR-1603: Update ManageEntitlements page to follow the changes on EntitlementCalculation
The ManageEntitlements page uses the EntitlementCalculator to create the instances of
EntitlementCalculation wich will be displayed.

The EntitlementCalculator was updated to created calculations based on the contact instead of
contract. The ManageEntitlements was updated in the same way. It now expects the cid
parameter to be a list of contacts ids and then uses the HRJobContract
getContactsWithActiveContracts API endpoint to fetch a list of contacts with active contracts
during the selected absence period. The page was also updated to use the contact id to index
the form fields instead of the contract id, which was used before.