PCHR-1607: Move to the next available Absence Period after saving entitlements
The postProcess() method was updated to check if there are other periods next
to the one being processed and, if so, is takes the entitlement calculation page
of the next period until there are no more period available. When that happens,
the user is taken back to the page where they was before starting the whole
entitlement calculation process.