PCHR-1607: Make getProposedEntitlement() to always return the proposed entitlement, even if it was overridden
Recently, the way to store Overridden entitlements has changed. Now, we store the
overidden value as the difference between the proposed entitlement and the overridden
value. So, instead of saving a single balance change with the overridden value, we store
one with the proposed value and another with the difference (overridden_value -
proposed_value).

The getProposedEntitlement() (used to save the balance_change with the proposed
entitlement) method was checking if there was a previously calculated LeavePeriodEntitlement
for the calculation's period and would return it's overridden value if it was overridden.
This would make you lose the proposed entitlement information if someone tried to
recalculate the entitlement for that period and set a new overridden value. Example:

You save the entitlement for a period one first time. The proposed entitlement is 20, but
you override it to 50. In that case, two LeaveBalanceChanges will be created: one with
the proposed entitlement (20) and another one with the difference between the overridden
value and the proposed one (30). Next, you try to recalculate the entitlement for that period,
but now the proposed entitlement will be 50 (the overridden value) instead of 20, which
would be the right calculated value.

This was fixed by changing the getProposedEntitlement() to never include the overridden
entitlement. Besides, a new method, getOverriddenEntitlement(), was created in order
to return a previously overridden entitlement, if it exists.

This also allowed one improvement to the Manage Entitlements page. Now, when using it
to recalculate the entitlements for a previously overridden entitlement, it will display the
overridden value but, if the user unchecks the "override" checkbox, the original proposed
value will be displayed. This was not possible before.